feat(move): ディレクトリ分けを詳細化
